+++ packages/gecko_db/test/phase0_3_defensive_coverage_test.dart
@@ -60,6 +60,9 @@ class _DelayedBackend implements RawBackend {
   Future<int> liveQueryCount() => delegate.liveQueryCount();
 
   @override
+  Future<List<RawEntry>> pendingChanges() => delegate.pendingChanges();
+
+  @override
   Future<RawSnapshot> snapshot() => delegate.snapshot();
 
   @override