+++ packages/gecko_db/test/phase2_bounds_test.dart
@@ -112,6 +112,9 @@ class _DelayedBackend implements RawBackend {
   Future<int> liveQueryCount() => delegate.liveQueryCount();
 
   @override
+  Future<List<RawEntry>> pendingChanges() => delegate.pendingChanges();
+
+  @override
   Future<RawSnapshot> snapshot() => delegate.snapshot();
 
   @override